style imports, rename folders and files, update userState thunks
