+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
+import "./App.css";
 import reactLogo from "./assets/react.svg";
 import viteLogo from "/logo.svg";
-import "./App.css";
 
 function App() {
   const [count, setCount] = useState(0);
