+++ src/main.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { Provider } from "react-redux";
+import { PersistGate } from "redux-persist/integration/react";
 import App from "./App.tsx";
 import "./index.css";
-import { Provider } from "react-redux";
-import { store } from "./store/store.ts";
-import { persistor } from "./store/store.ts";
-import { PersistGate } from "redux-persist/integration/react";
+import { persistor, store } from "./store/store.ts";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
