+++ src/types/index.ts
@@ -1,3 +1,2 @@
-export * from "./login";
 export * from "./User";
 export * from "./Token";
