--- src/types/login.ts
@@ -1,4 +1,0 @@
-export type Login = {
-  email: string;
-  password: string;
-};
